+++ commands/spigot-platform/src/main/java/net/apartium/cocoabeans/commands/spigot/requirements/factory/PermissionFactory.java
@@ -78,14 +78,12 @@ public class PermissionFactory implements RequirementFactory {
 
         private class UnmetPermissionResponse extends UnmetRequirementResponse {
 
-
-
             public UnmetPermissionResponse(Requirement requirement, RequirementEvaluationContext context, String message) {
                 super(requirement, context, message, permission);
             }
 
             @Override
-            public Exception getError() {
+            public PermissionException getError() {
                 return new PermissionException(this, permission);
             }
         }